Added permissions to task detail page

+++ src/pages/api/projects/[id]/tasks/[taskId]/index.ts
@@ -1,23 +1,91 @@
+import { permissions } from '../../../../../../config';
 import {
 	prisma,
+	getProject,
 	getProjectTask,
 	taskSelectQuery as selectQuery,
 } from '../../../../../../db';
+import {
+	getRecord,
+	getUserObjectPermissions,
+	hasViewPermission,
+	removeObjectPermissions,
+	updateObjectPermissions,
+} from '../../../../../../db/utils';
 import { auth } from '../../../../../../middlewares';
-import { CreateProjectTaskQueryType } from '../../../../../../types';
+import {
+	CreateProjectTaskQueryType,
+	ProjectTaskType,
+} from '../../../../../../types';
+import { hasModelPermission } from '../../../../../../utils';
+import { NextApiErrorMessage } from '../../../../../../utils/classes';
 import { taskCreateSchema } from '../../../../../../validators';
 
 export default auth()
+	.use(async (req, res, next) => {
+		// Check the user can view the project
+		const canViewProject = await hasViewPermission({
+			model: 'projects',
+			perm: 'project',
+			objectId: req.query.id as string,
+			user: req.user,
+		});
+		if (!canViewProject) throw new NextApiErrorMessage(403);
+		next();
+	})
 	.get(async (req, res) => {
-		const task = await getProjectTask(req.query.taskId as string);
+		const record = await getRecord({
+			model: 'projects_tasks',
+			perm: 'projecttask',
+			permission: 'VIEW',
+			objectId: req.query.id as string,
+			user: req.user,
+			getData() {
+				return getProjectTask(req.query.taskId as string);
+			},
+		});
+
+		if (!record) throw new NextApiErrorMessage(403);
+
+		if (!record.data) {
+			return res.status(404).json({
+				status: 'error',
+				message: 'Task with specified ID does not exist!',
+			});
+		}
 
 		return res.status(200).json({
 			status: 'success',
-			message: 'Fetched Project Task Successfully!',
-			data: task,
+			message: 'Fetched task successfully!',
+			data: record.data,
 		});
 	})
 	.put(async (req, res) => {
+		let hasPerm =
+			req.user.isSuperUser ||
+			hasModelPermission(req.user.allPermissions, [
+				permissions.projecttask.EDIT,
+			]);
+
+		if (!hasPerm) {
+			const perm = await getUserObjectPermissions({
+				modelName: 'projects_tasks',
+				permission: 'EDIT',
+				objectId: req.query.id as string,
+				userId: req.user.id,
+			});
+			hasPerm = perm.edit;
+		}
+
+		if (!hasPerm) throw new NextApiErrorMessage(403);
+
+		const previousTask = await getProjectTask(req.query.id as string);
+		if (!previousTask)
+			return res.status(404).json({
+				status: 'error',
+				message: 'Task with the specified ID does not exist',
+			});
+
 		const valid: CreateProjectTaskQueryType =
 			await taskCreateSchema.validateAsync({ ...req.body });
 
@@ -25,12 +93,22 @@ export default auth()
 
 		// delete old project task followers in a team array is passed
 		if (followers && Array.isArray(followers))
-			await prisma.projectTaskFollower.deleteMany({
-				where: { taskId: req.query.taskId as string },
-			});
+			await Promise.all([
+				prisma.projectTaskFollower.deleteMany({
+					where: { taskId: req.query.taskId as string },
+				}),
+				// Also remove their user id from the project taskobject permissions
+				await removeObjectPermissions({
+					model: 'projects_tasks',
+					objectId: previousTask.id,
+					users: previousTask.followers.map(
+						(follower) => follower.member.employee.user.id
+					),
+				}),
+			]);
 
 		// update the project task
-		const project = await prisma.projectTask.update({
+		const task = (await prisma.projectTask.update({
 			where: { id: req.query.taskId as string },
 			data: {
 				...data,
@@ -48,21 +126,60 @@ export default auth()
 						: {},
 			},
 			select: selectQuery,
-		});
+		})) as unknown as ProjectTaskType;
+
+		const taskFollowers = task.followers.map(
+			(follower) => follower.member.employee.user.id
+		);
+
+		await Promise.all([
+			updateObjectPermissions({
+				model: 'projects_tasks',
+				permissions: ['VIEW'],
+				objectId: task.id,
+				users: taskFollowers,
+			}),
+			...task.followers.map((follower) =>
+				updateObjectPermissions({
+					model: 'projects_tasks_followers',
+					permissions: ['VIEW'],
+					objectId: follower.id,
+					users: taskFollowers,
+				})
+			),
+		]);
 
 		return res.status(200).json({
 			status: 'success',
-			message: 'Project Task was updated successfully',
-			data: project,
+			message: 'Task was updated successfully',
+			data: task,
 		});
 	})
 	.delete(async (req, res) => {
+		let hasPerm =
+			req.user.isSuperUser ||
+			hasModelPermission(req.user.allPermissions, [
+				permissions.projecttask.DELETE,
+			]);
+
+		if (!hasPerm) {
+			const perm = await getUserObjectPermissions({
+				modelName: 'projects_tasks',
+				permission: 'DELETE',
+				objectId: req.query.id as string,
+				userId: req.user.id,
+			});
+			hasPerm = perm.delete;
+		}
+
+		if (!hasPerm) throw new NextApiErrorMessage(403);
+
 		await prisma.projectTask.delete({
 			where: { id: req.query.taskId as string },
 		});
 
 		return res.status(200).json({
 			status: 'success',
-			message: 'Project task deleted successfully!',
+			message: 'Task deleted successfully!',
 		});
 	});
